Download_videos/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
decorator==4.4.2
Django==4.2.5
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
greenlet==2.0.2
idna==3.4
imageio==2.31.3
imageio-ffmpeg==0.4.8
moviepy==1.0.3
numpy==1.25.2
Pillow==10.0.0
playwright==1.37.0
proglog==0.1.10
pyee==9.0.4
PyJWT==2.9.0
pytube==15.0.0
pytube3==9.6.4
pytubefix==8.12.2
requests==2.31.0
sqlparse==0.4.4
TikTokApi==6.1.1
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
yt-dlp==2025.1.15
